The rev5 board design proposal.
Removed the charge pump.
When the common pin for both LEDs is driven with a 50% duty-cycle rectangular signal both LEDs can be used without additional circuit.
QMK has a feature to change the brightness of background LEDs and we are just going to use this feature to produce such a signal.

As the FN switch is also connected to this line we have to filter out the switching frequency to be able to read the switch.
For this a simple diode rectifier was added. A resistor to discharge the filter capacitor is also needed to detect the switch off state.